GerritURL implemented; ViewPager now refreshes tabs on reloading.
The ViewPager still is caching tabs, and not redrawing the ones that are already loaded. However, it cannot have all the tabs loaded at once so with enough tab switching they will be up to date. There is (still) problems with tabs showing the wrong data (projects not being set/unset, wrong status such as merged instead of abandoned etc.). We want to force a redraw of the fragment's layout on refresh instead of tearing it down and recreating it.

With the GerritURL class, we can set individual components of the query url. This is especially when components change.

Use a FragmentStatePagerAdapter for the ViewPager and manage the tabs inside this class. This manages the tab titles and resets (recreates) one when the title changes.
Set the project globally from the preferences and add a listener for project changes.
CardsUI (library): Refresh invalidates the parent layout
GerritURL: Better string construction
CardsFragment: Remove precautionary GerritTask management as this should be done by an activity instead.
TrackingClickListener: Remove intents to startActivity due to use of fragments.
MyTabListener: Import re-ordering