Fixed some issues with new project list
Moved the adapter and loader out into seperate files
Removed the original project listing and replaced with new one
Fix some string parsing issues
Get the root project when a child project is selected

This new project list is UNSTABLE as there are a few issues to be fixed. Currently, opening the project list does not display anythinng.