Fix loading selected change details
Don't reuse the bundle when instantiating each page of the change details screen (phones)
as the last page the viewpager instantiates is not always the first page (the first change in the
list will always be brought up first and the fragment arguments to load the change we want
will be overwritten.
Fix version number comparison - 2.10 > 2.8. AOKP will now not use the legacy change details
Default diff options to internal viewer
Bug fix for getting change commit number
Reformat README and TODO list

Change-Id: I8360394fd39642002add37e733fec31b9f4a92a0